populate script user input is now a data utility
