+++ src/scripts/populate-cve.js
@@ -3,16 +3,19 @@
  * collection
  */
 
-const express = require('express')
 const config = require('config')
-const mongoose = require('mongoose')
+
+const express = require('express')
 const app = express()
+const mongoose = require('mongoose')
+
+const dataUtils = require('../utils/data')
+const errors = require('../utils/error')
 const logger = require('../middleware/logger')
 const Cve = require('../model/cve')
-const cveData = require('../../datadump/pre-population/cves')
-const ProgressBar = require('progress')
-const errors = require('../utils/error')
+
 const error = new errors.IDRError()
+
 require('dotenv').config() // This enables us to read from the .env file.
 
 // Body Parser Middleware
@@ -57,23 +60,10 @@ db.on('error', () => {
 })
 
 db.once('open', async () => {
-  // we're connected!
   logger.info(`Successfully connected to database ${dbName} at ${dbHost}:${dbPort}`)
 
-  // Ask user to confirm pre-population, which will prep MongoDB by removing the Cve-Id, Org, and User collections
-  const prompt = require('prompt-sync')({ sigint: true })
-  let userInput = prompt(
-    `Are you sure you wish to pre-populate the database for the ${appEnv} environment?` +
-    `Doing so will drop the 'Cve' collection in the ${dbName} database. (y/n) `
-  )
-
-  while (userInput.toLowerCase() !== 'n' && userInput.toLowerCase() !== 'y') {
-    console.log('Unrecognized Input')
-    userInput = prompt(
-      `Do you wish to pre-populate the database for the ${appEnv} environment?` +
-      `Doing so will drop the 'Cve', collection in the ${dbName} database. (y/n) `
-    )
-  }
+  // user needs to agree to an action that drops collections
+  const userInput = dataUtils.getUserPopulateInput()
 
   // dropping Cve collection
   // this may be more complex than necessary, but is kept similar to `src/scripts/populate.js`
@@ -95,39 +85,19 @@ db.once('open', async () => {
     })
 
     if (!names.includes('Cve')) {
-      await populateCveCollection()
-      logger.info('Successfully pre-populated the database!')
+      await dataUtils.populateCollection(
+        './datadump/pre-population/cves.json',
+        Cve
+      )
+      logger.info('Successfully populated CVE records!')
     } else {
-      logger.error('The database was not populated because some of the collections were not deleted.')
+      logger.error(
+        'The database was not populated because ' +
+        'some of the collections were not deleted.'
+      )
     }
   }
 
-  // close MongoDB connection
   mongoose.connection.close()
 })
 
-// Populating Cve collection
-async function populateCveCollection () {
-  const bar = new ProgressBar('Cve collection [:bar] :percent', {
-    complete: '=',
-    incomplete: ' ',
-    width: 40,
-    total: cveData.length
-  })
-
-  let batchCves = []
-  for (let i = 0; i < cveData.length; i++) {
-    const cve = cveData[i]
-    batchCves.push(cve)
-
-    const isLastItem = i === cveData.length - 1
-    if (i % 1000 === 0 || isLastItem) {
-      await Cve.insertMany(batchCves)
-      bar.tick(batchCves.length)
-      batchCves = []
-    }
-  }
-
-  // success!
-  logger.info('Cve collection populated')
-}
