+++ src/scripts/populate.js
@@ -4,6 +4,7 @@
  */
 
 const config = require('config')
+
 const express = require('express')
 const app = express()
 const mongoose = require('mongoose')
@@ -72,24 +73,8 @@ db.on('error', () => {
 db.once('open', async () => {
   logger.info(`Successfully connected to database ${dbName} at ${dbHost}:${dbPort}`)
 
-  // Ask user to confirm pre-population
-  // This will prep MongoDB by removing the Cve, Cve-Id-Range, Cve-Id,
-  // Org, and User collections
-  /*const prompt = require('prompt-sync')({ sigint: true })
-  let userInput = prompt(
-    `Are you sure you wish to pre-populate the database for the ${appEnv} environment? ` +
-    "Doing so will drop the 'Cve', 'Cve-Id-Range', 'Cve-Id', 'Org' and 'User' collections " +
-    `in the ${dbName} database. (y/n) `
-  )
-  while (userInput.toLowerCase() !== 'n' && userInput.toLowerCase() !== 'y') {
-    console.log('Unrecognized Input')
-    userInput = prompt(
-      `Do you wish to pre-populate the database for the ${appEnv} environment? ` +
-      "Doing so will drop the 'Cve', 'Cve-Id-Range', 'Cve-Id', 'Org' and 'User' collections " +
-      `in the ${dbName} database. (y/n) `
-    )
-  }*/
-  let userInput = 'y'
+  // script runner (currently) needs to agree to an action that drops collections
+  const userInput = dataUtils.getUserPopulateInput()
 
   // drops and re-populates collections
   if (userInput.toLowerCase() === 'y') {
@@ -157,7 +142,9 @@ db.once('open', async () => {
     } else {
       logger.error(
         'The database was not populated because ' +
-        'some of the collections were not deleted.')
+        'some of the collections were not deleted.'
+      )
+      mongoose.connection.close()
     }
   }
 })
