+++ src/utils/data.js
@@ -11,6 +11,7 @@ const color = require('kleur')
 const cryptoRandomString = require('crypto-random-string')
 const JSONStream = require('JSONStream')
 const mongoose = require('mongoose')
+const prompt = require('prompt-sync')({ sigint: true })
 const uuid = require('uuid')
 
 const CONSTANTS = require('../constants')
@@ -90,6 +91,21 @@ async function newCveIdTransform (cveId) {
   return cveId
 }
 
+function getUserPopulateInput (appName, dbName) {
+  let userInput = prompt(
+    `Are you sure you wish to pre-populate the database for the ${appName} environment?` +
+    `Doing so will drop the 'Cve' collection in the ${dbName} database. (y/n) `
+  )
+  while (userInput.toLowerCase() !== 'n' && userInput.toLowerCase() !== 'y') {
+    console.log('Unrecognized Input')
+    userInput = prompt(
+      `Do you wish to pre-populate the database for the ${appName} environment?` +
+      `Doing so will drop the 'Cve', collection in the ${dbName} database. (y/n) `
+    )
+  }
+  return userInput
+}
+
 /* populates any collection given the file path for an input JSON,
  * the data model for the collection, a function that transforms the
  * data, and a hash that only gets use in the User collection
@@ -129,6 +145,7 @@ function promisesShifter (promises) {
 }
 
 module.exports = {
+  getUserPopulateInput,
   newCveIdTransform,
   newOrgTransform,
   newUserTransform,
